split out meta-producer resources
